FROMLIST: mtd: spi-nor: fix wrong "fully unlocked" test
In stm_unlock(), the test to determine whether we've fully unlocked the
flash checks for the lock length to be equal to the flash size. That is
a typo/think-o -- the condition actually means the flash is completely
*locked.* We should be using the inverse condition -- that the lock
length is 0 (i.e., no protection).

The result of this bug is that we never actually turn off the Status
Register Write Disable bit, even if the flash is completely unlocked.
Now we can.

Fixes: 47b8edbf0d43 ("mtd: spi-nor: disallow further writes to SR if WP# is low")
Reported-by: Giorgio <giorgio.nicole@arcor.de>
Signed-off-by: Brian Norris <computersforpeace@gmail.com>
Cc: Ezequiel Garcia <ezequiel@vanguardiasur.com.ar>
(am from https://patchwork.ozlabs.org/patch/640333/)

BUG=chrome-os-partner:54040, chrome-os-partner:54736
TEST=`flashrom --wp-XXX` commands; also mtdinfo, flash_lock, flash_unlock

Change-Id: Idc7e4736d80330e92b01ce7afc310f0eab42bab5
Signed-off-by: Brian Norris <briannorris@chromium.org>
Previous-Reviewed-on: https://chromium-review.googlesource.com/358162
(cherry picked from commit c30687792eff4986ddd011bff67f388ae81943e0)
Reviewed-on: https://chromium-review.googlesource.com/362176