CHROMIUM: mac80211: add packet trace debugging feature
The packet trace feature is to generate additional logs along
the data path, to capture the important tx and rx activities on
specified stations.

It creates two debugfs controls for each phyX interface under
ieee80211/phyX folder. The debugfs controls allow user to add or
remove MAC addresses to be traced, configure the types of frames
being traced, and enable or disable the feature itself.

This feature is disabled by default. Once configured and enabled,
it identifies the frames to be sent or just received, and traces
these frames along the data path in both transmit and receive
direction. Whenever these frames are dropped in various functions
logs will be generated.

Tracing information is carried in TXCB and RXCB in sk_buff. APIs
are provided so drivers supporting mac80211 can use this feature
to provide more insight of tx and rx activities.

BUG=none
TEST=manually tested feature functionality on whirlwind.
     1. Tested commands to enable/disable.
        echo 1 > <debugfs>/ieee80211/phy1/packet_trace/ptenable
     2. Tested commands to configure with valid and invalid input.
     3. Manually add station MAC, associate station, check log.
        echo add sta <MAC> .../phy1/packet_trace/ptctrl

Change-Id: I0875a2b5fdc1753dc054f16238ac375776685c71
Reviewed-on: https://chromium-review.googlesource.com/353333
Commit-Ready: Zhihong Yu <zhihongyu@chromium.org>
Tested-by: Zhihong Yu <zhihongyu@chromium.org>
Reviewed-by: Srinivasa duvvuri <sduvvuri@chromium.org>