CHROMIUM: iwl7000: mvm: support v4 of the TX power command
Add support for the v4 version of the TX power command.  Just add a
new version and do the same sizing tricks that were done when support
for v3 was introduced.

This patch doesn't support the new functionality introduced, but makes
the driver work with the new size of the command.

Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 3a596896bcbe48683139aeed6882432bb0b9d129