UPSTREAM: mmc: sdhci: Record what command is using the data lines
In order to support commands during data transfer, there must be a
distinction between the command that is using the command line (and
for which a command interrupt is expected) and the command that is
using the data lines (for which a data interrupt is expected).

There is host->cmd for the command line, but there is only host->data
for the data lines, which is a different structure, does not represent
the command in use, and is anyway NULL in the case of commands that use
the data lines for busy signalling instead of data transfer.

Introduce host->data_cmd to record what command is using the data lines,
and use that instead of host->cmd when referring to the data command.

BUG=b:117176477, b:123303213
TEST=To sample the responses of cmd in the failing window to see if the
     kernel can detect the interrupt of ACMD23 CRC error.

Change-Id: I6b274b8dded9c04f4d6ed16d430b2815ab5a5347
Signed-off-by: Adrian Hunter <adrian.hunter@intel.com>
Signed-off-by: Ulf Hansson <ulf.hansson@linaro.org>
(cherry picked from commit 7c89a3d9082c316f898ec2c3c6c1689e0cad6117)
Signed-off-by: Kaiyen Chang <kaiyen.chang@intel.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/1405963
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Nelson Ye <nelson_ye@compal.corp-partner.google.com>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@google.com>
(cherry picked from commit 80af01703fabda44dd9310b2a47d956a36fc5852)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/1541842
Commit-Queue: Gwendal Grignou <gwendal@google.com>
Tested-by: Gwendal Grignou <gwendal@google.com>