BACKPORT: nl80211: support beacon report scanning
Beacon report radio measurement requires reporting observed BSSs
on the channels specified in the beacon request. If the measurement
mode is set to passive or active, it requires actually performing a
scan (passive or active, accordingly), and reporting the time that
the scan was started and the time each beacon/probe was received
(both in terms of TSF of the BSS of the requesting AP). If the
request mode is table, this information is optional.
In addition, the radio measurement request specifies the channel
dwell time for the measurement.

In order to use scan for beacon report when the mode is active or
passive, add a parameter to scan request that specifies the
channel dwell time, and add scan start time and beacon received time
to scan results information.

Supporting beacon report is required for Multi Band Operation (MBO).

Signed-off-by: Assaf Krauss <assaf.krauss@intel.com>
Signed-off-by: David Spinadel <david.spinadel@intel.com>
Signed-off-by: Avraham Stern <avraham.stern@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
(cherry picked from commit 1d76250bd34af86c6498fc51e50cab3bfbbeceaa)

 Conflicts:
	drivers/net/wireless/marvell/mwifiex/cmdevt.c
	drivers/staging/wilc1000/wilc_wfi_cfgoperations.c
        (manual addition) drivers/net/wireless/ar10k/mac80211/scan.c

kirtika@: Manual wiggling to populate struct cfg80211_scan_info
before a cfg80211_scan_done calls in both these drivers, without
changing anything else.

BUG=chromium:828485
TEST=build on eve and kevin

Change-Id: I0f4f8f4bf3e4fca5c4c81ed6ff862bc6a835a477
Reviewed-on: https://chromium-review.googlesource.com/1004128
Commit-Ready: Kirtika Ruchandani <kirtika@chromium.org>
Tested-by: Kirtika Ruchandani <kirtika@chromium.org>
Reviewed-by: Kirtika Ruchandani <kirtika@chromium.org>