BACKPORT: KVM: nVMX: keep preemption timer enabled during L2 execution
Because the vmcs12 preemption timer is emulated through a separate hrtimer,
we can keep on using the preemption timer in the vmcs02 to emulare L1's
TSC deadline timer.

However, the corresponding bit in the pin-based execution control field
must be kept consistent between vmcs01 and vmcs02.  On vmentry we copy
it into the vmcs02; on vmexit the preemption timer must be disabled in
the vmcs01 if a preemption timer vmexit happened while in guest mode.

The preemption timer value in the vmcs02 is set by vmx_vcpu_run, so it
need not be considered in prepare_vmcs02.

Cc: Yunhong Jiang <yunhong.jiang@intel.com>
Cc: Haozhong Zhang <haozhong.zhang@intel.com>
Tested-by: Wanpeng Li <kernellwp@gmail.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>

BUG=b:120040011
TEST=Built

(cherry picked from commit 9314006db8b781715658cd6a28994d84ccce5dee)
Signed-off-by: Matt Delco <delco@chromium.org>
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Change-Id: I1f8084a8ca16a7add6b5f196ddcaf10dbcf1a16d
Reviewed-on: https://chromium-review.googlesource.com/1766518
Legacy-Commit-Queue: Commit Bot <commit-bot@chromium.org>