CHROMIUM: drivers: nss-ecm:  Update accel engine interface id
In some cases, the accel engine interface identifier
can be changed. In these cases, when the new connections
are being created, ECM does not establish a new interface
instance, because there is already one in the database with
the same MAC address and netdev index number. But since the
accel engine interface identifier is changed, we need to update
the interface instances ae_interface_identifier field.

BUG=chrome-os-partner:54309
TEST=build/boot on WW. Verified with mesh create steps listed in the bug
and confirm the interface identifier is updated correctly.

Change-Id: Id60031fca831f29a2e84db3ea12bf2875441b3df
Signed-off-by: Stephen Wang <wstephen@codeaurora.org>
Reviewed-on: https://chromium-review.googlesource.com/358551
Commit-Ready: Srinivasa duvvuri <sduvvuri@chromium.org>
Tested-by: Kishan Kunduru <kkunduru@chromium.org>
Reviewed-by: Kevin Hayes <kevinhayes@google.com>
Reviewed-by: Kishan Kunduru <kkunduru@chromium.org>