CHROMIUM: iwl7000: pcie: merge iwl_queue and iwl_txq
The original intent was to have the general iwl_queue shared
between RX and TX queues, but it is not the actual status.
Since it is not shared with any struct but iwl_txq, it adds
unnecessary complexity. Merge those structs.

Signed-off-by: Sara Sharon <sara.sharon@intel.com>
iwl7000-tree: d7ea515e3eace0323948dcdbbef131c58fa80534