WIP: input: cros_ec_keyb: Add non-matrix buttons and switches
On some newer boards using mkbp we're hooking up non-matrix buttons and
switches to the EC but NOT to the main application processor.  The EC
has recently had some new commands added for allowing the kernel to work
with these buttons and switches, namely:
- http://crosreview.com/358633
- http://crosreview.com/358634
- http://crosreview.com/358926

Let's add kernel support to handle this.  Rather than creating a whole
new input driver, we'll continue to use cros_ec_keyb and just report the
new keys.

BUG=chrome-os-partner:54988
BUG=chrome-os-partner:54976
TEST=Try lid, power, and volume on gru w/ evtest
TEST=Boot lid closed: evtest --query /dev/*/event0 EV_SW SW_LID; echo $?
TEST=Suspend/resume w/ lid while running evtest
TEST=Try on minnie; no trouble found

Change-Id: Id57f1e80ee9772b0fc5fc3fdd2a5993e9a45f31b
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/362366