UPSTREAM: ACPI / scan: fix enumeration (visited) flags for bus rescans
If the ACPI tables change as a result of a dinamically loaded table
and a bus rescan is required the enumeration/visited flag are not
consistent.

I2C/SPI are not directly enumerated in acpi_bus_attach(), however
the visited flag is set. This makes it impossible to check if an
ACPI device has already been enumerated by the I2C and SPI
subsystems. To fix this issue we only set the visited flags if
the device is not I2C or SPI.

With this change we also need to remove setting visited to false
from acpi_bus_attach(), otherwise if we rescan already enumerated
I2C/SPI devices we try to re-enumerate them.

Note that I2C/SPI devices can be enumerated either via a scan handler
(when using PRP0001) or via regular device_attach(). In either case
the flow goes through acpi_default_enumeration() which makes it the
ideal place to mark the ACPI device as enumerated.

Signed-off-by: Octavian Purdila <octavian.purdila@intel.com>
Reviewed-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
(cherry picked from commit 10c7e20b2ff3caa5a8c0e7d60aef5a9c86e60ce8)
Signed-off-by: Mike Mason <michael.w.mason@intel.com>

BUG=b:78601949
TEST=Rebuild/install coreboot with CL 1302135 per comment #93 (b/78601949)
     Build kernel and install on a DUT with a 1 MHz i2c slave device.
     Review dmesg to verify I2C device is detected and not reporting errors.
     Operate device (ie touch trackpad) to verify correct operation.
     Review dmesg to verify correct I2C parameters are used. e.g.:

[    0.496636] ... i2c_designware.1: clk_freq 1000000, acpi_speed 1000000
...
[    0.496656] dw_i2c_acpi_cfg: FPCN 25 4a 24
^^^ DEBUG output provided by CL 1085476 ^^^
...
[    0.767995] input: ACPI0C50:00 06CB:1058 as
/devices/pci0000:00/0000:00:15.0/i2c_designware.0/i2c-6/i2c-ACPI0C50:00/0018:06CB:1058.0001/input/input3
[    0.768251] hid-multitouch 0018:06CB:1058.0001: input,hidraw0: I2C HID v1.00 Device [ACPI0C50:00 06CB:1058] on i2c-ACPI0C50:00
[    0.768297] i2c_hid i2c-ACPI0C50:01: GPIO lookup for consumer reset
[    0.768300] i2c_hid i2c-ACPI0C50:01: using ACPI for GPIO lookup
[    0.768303] acpi ACPI0C50:01: GPIO: looking up reset-gpios
[    0.768306] acpi ACPI0C50:01: GPIO: looking up reset-gpio
[    0.768308] acpi ACPI0C50:01: GPIO: looking up 0 in _CRS
[    0.768321] i2c_hid i2c-ACPI0C50:01: using lookup tables for GPIO lookup
[    0.768324] i2c_hid i2c-ACPI0C50:01: lookup for GPIO reset failed
[    0.806943] input: ACPI0C50:01 04F3:30C5 Touchpad as
/devices/pci0000:00/0000:00:15.2/i2c_designware.1/i2c-7/i2c-ACPI0C50:01/0018:04F3:30C5.0002/input/input5
[    0.807026] hid-multitouch 0018:04F3:30C5.0002: input,hidraw1: I2C HID v1.00 Mouse [ACPI0C50:01 04F3:30C5] on i2c-ACPI0C50:01

Change-Id: I1a22f98e7df07d8db3c5d4a8ca64f500d03938b0
Reviewed-on: https://chromium-review.googlesource.com/1085385
Commit-Ready: Grant Grundler <grundler@chromium.org>
Tested-by: Grant Grundler <grundler@chromium.org>
Reviewed-by: Grant Grundler <grundler@chromium.org>