CHROMIUM: sysrq: replay the correct keys when releasing SysRq combo
When we changed SysRq code to terminate the SysRq state when releasing
either Alt or SysRq (upstream allows releasing SysRq and only needing to
keep pressing Alt) we neglected to adjust the event re-inject code which
caused us to keep re-injecting Alt-SysRq combination every time SysRq
alone was pressed until user again pressed and released Alt key.

Also we allowed F10 (Volume UP) to stand in for SysRq (since we do not
have dedicated SysRq key on our keyboards) but re-inject code was not
aware of this.

BUG=chromium:628508
TEST=1) Press Alt+Volume_Up (F10)
     2) Release both
     3) Press Volume_Up - OSD should show increased volume

Change-Id: I5ddb2e6fe3a85a2cf030697cc9a5467f6b0ca237
Signed-off-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/361554
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>